Automap levels can now be browsed up and down. where no automap exists, a blank automap is generated.

+++ src/base/UWClass.cs
@@ -5,14 +5,14 @@ namespace Underworld
     /// </summary>
     public class UWClass
     {
-        public const short GAME_UWDEMO = 0;//"UW0";
-        public const short GAME_UW1 = 1; // "UW1";
-        public const short GAME_UW2 = 2; // "UW2";
+        public const int GAME_UWDEMO = 0;//"UW0";
+        public const int GAME_UW1 = 1; // "UW1";
+        public const int GAME_UW2 = 2; // "UW2";
 
         /// <summary>
         /// Use to track what game is currently active.
         /// </summary>
-        public static short _RES = 1;  //"UW1";
+        public static int _RES = 1;  //"UW1";
 
         public static string BasePath;// = "C:\\Games\\UW2\\game\\UW2";
